consolidation
before, most of the information was done in separate .txt files, but that means multiple fetches - so i've consolidated most of these to single .json file. This mostly concerns the about page. now i use a single about.json that handles all that information. quicker and more reliable loading without the need to parse the data.

i originally didnt have a single .js for the about page like i do the other stuff, it was harder to fix issues so i've consolidated it all into one.